WNN High-Integrity Final Sweep & Copilot Resolutions

+++ src/calibration.hpp
@@ -16,7 +16,7 @@ public:
     bool run_sweep(long double& out_omega);
 
     // Persistence Layer
-    static bool save_profile(const std::string& path, double omega, double alpha, double beta, double delta);
+    static bool save_profile(const std::string& path, long double omega, long double alpha, long double beta, long double delta);
     static bool load_profile(const std::string& path, long double& omega, long double& alpha, long double& beta, long double& delta);
 };
